refactor : ID 데이터 타입  변경
- Board, Laundry, User Id 데이터 타입 int에서 Long 으로 변경
- Long이 JPA ID값 권장사항, null과 0을 따로 구분하기 위해
- 해당 ID값을 수정하여 관련된 데이터타입 모두 변경

+++ Server/wiselaundry/src/main/java/com/ssafy/wiselaundry/domain/board/response/BoardCreateRes.java
@@ -13,9 +13,9 @@ import lombok.ToString;
 @ApiModel("생성된 게시판 ID 리턴")
 public class BoardCreateRes extends BaseResponseBody {
     @ApiModelProperty(value = "게시판 ID", required = true)
-    private int boardId;
+    private long boardId;
 
-    public static BoardCreateRes of (Integer statusCode, String message, Integer boardId){
+    public static BoardCreateRes of (Integer statusCode, String message, long boardId){
         BoardCreateRes res = new BoardCreateRes();
         res.setStatusCode(statusCode);
         res.setMessage(message);
